Move more nested classes out
