+++ Terminal.Gui/Views/TableView/ListColumnStyle.cs
@@ -11,7 +11,7 @@
 		public Orientation Orientation { get; set; } = Orientation.Horizontal;
 
 		/// <summary>
-		/// Gets or sets a flag indicating whether to scroll in the same direction as <see cref="ListTableSource.ListColumnStyle.Orientation"/>.
+		/// Gets or sets a flag indicating whether to scroll in the same direction as <see cref="ListColumnStyle.Orientation"/>.
 		/// Defaults to <see langword="false"/>.
 		/// </summary>
 		public bool ScrollParallel { get; set; } = false;
