+++ Terminal.Gui/Views/TableView/TableView.cs
@@ -4,9 +4,22 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Data;
 using System.Linq;
-using static Terminal.Gui.SpinnerStyle;
 
 namespace Terminal.Gui {
+
+	/// <summary>
+	/// Delegate for providing color to <see cref="TableView"/> cells based on the value being rendered
+	/// </summary>
+	/// <param name="args">Contains information about the cell for which color is needed</param>
+	/// <returns></returns>
+	public delegate ColorScheme CellColorGetterDelegate (CellColorGetterArgs args);
+
+	/// <summary>
+	/// Delegate for providing color for a whole row of a <see cref="TableView"/>
+	/// </summary>
+	/// <param name="args"></param>
+	/// <returns></returns>
+	public delegate ColorScheme RowColorGetterDelegate (RowColorGetterArgs args);
 
 	/// <summary>
 	/// View for tabular data based on a <see cref="ITableSource"/>.
@@ -1767,248 +1780,7 @@ namespace Terminal.Gui {
 			return colStyle != null ? colStyle.GetRepresentation (value) : value.ToString ();
 		}
 
-		/// <summary>
-		/// Delegate for providing color to <see cref="TableView"/> cells based on the value being rendered
-		/// </summary>
-		/// <param name="args">Contains information about the cell for which color is needed</param>
-		/// <returns></returns>
-		public delegate ColorScheme CellColorGetterDelegate (CellColorGetterArgs args);
-
-		/// <summary>
-		/// Delegate for providing color for a whole row of a <see cref="TableView"/>
-		/// </summary>
-		/// <param name="args"></param>
-		/// <returns></returns>
-		public delegate ColorScheme RowColorGetterDelegate (RowColorGetterArgs args);
-
-		#region Nested Types
-		/// <summary>
-		/// Describes how to render a given column in  a <see cref="TableView"/> including <see cref="Alignment"/> 
-		/// and textual representation of cells (e.g. date formats)
-		/// 
-		/// <a href="https://gui-cs.github.io/Terminal.Gui/articles/tableview.html">See TableView Deep Dive for more information</a>.
-		/// </summary>
-		public class ColumnStyle {
-
-			/// <summary>
-			/// Defines the default alignment for all values rendered in this column.  For custom alignment based on cell contents use <see cref="AlignmentGetter"/>.
-			/// </summary>
-			public TextAlignment Alignment { get; set; }
-
-			/// <summary>
-			/// Defines a delegate for returning custom alignment per cell based on cell values.  When specified this will override <see cref="Alignment"/>
-			/// </summary>
-			public Func<object, TextAlignment> AlignmentGetter;
-
-			/// <summary>
-			/// Defines a delegate for returning custom representations of cell values.  If not set then <see cref="object.ToString()"/> is used.  Return values from your delegate may be truncated e.g. based on <see cref="MaxWidth"/>
-			/// </summary>
-			public Func<object, string> RepresentationGetter;
-
-			/// <summary>
-			/// Defines a delegate for returning a custom color scheme per cell based on cell values.
-			/// Return null for the default
-			/// </summary>
-			public CellColorGetterDelegate ColorGetter;
-			private bool visible = true;
-
-			/// <summary>
-			/// Defines the format for values e.g. "yyyy-MM-dd" for dates
-			/// </summary>
-			public string Format { get; set; }
-
-			/// <summary>
-			/// Set the maximum width of the column in characters.  This value will be ignored if more than the tables <see cref="TableView.MaxCellWidth"/>.  Defaults to <see cref="TableView.DefaultMaxCellWidth"/>
-			/// </summary>
-			public int MaxWidth { get; set; } = TableView.DefaultMaxCellWidth;
-
-			/// <summary>
-			/// Set the minimum width of the column in characters.  Setting this will ensure that
-			/// even when a column has short content/header it still fills a given width of the control.
-			/// 
-			/// <para>This value will be ignored if more than the tables <see cref="TableView.MaxCellWidth"/> 
-			/// or the <see cref="MaxWidth"/>
-			/// </para>
-			/// <remarks>
-			/// For setting a flexible column width (down to a lower limit) use <see cref="MinAcceptableWidth"/>
-			/// instead
-			/// </remarks>
-			/// </summary>
-			public int MinWidth { get; set; }
-
-			/// <summary>
-			/// Enables flexible sizing of this column based on available screen space to render into.
-			/// </summary>
-			public int MinAcceptableWidth { get; set; } = DefaultMinAcceptableWidth;
-
-			/// <summary>
-			/// Gets or Sets a value indicating whether the column should be visible to the user.
-			/// This affects both whether it is rendered and whether it can be selected. Defaults to
-			/// true.
-			/// </summary>
-			/// <remarks>If <see cref="MaxWidth"/> is 0 then <see cref="Visible"/> will always return false.</remarks>
-			public bool Visible { get => MaxWidth >= 0 && visible; set => visible = value; }
-
-			/// <summary>
-			/// Returns the alignment for the cell based on <paramref name="cellValue"/> and <see cref="AlignmentGetter"/>/<see cref="Alignment"/>
-			/// </summary>
-			/// <param name="cellValue"></param>
-			/// <returns></returns>
-			public TextAlignment GetAlignment (object cellValue)
-			{
-				if (AlignmentGetter != null)
-					return AlignmentGetter (cellValue);
-
-				return Alignment;
-			}
-
-			/// <summary>
-			/// Returns the full string to render (which may be truncated if too long) that the current style says best represents the given <paramref name="value"/>
-			/// </summary>
-			/// <param name="value"></param>
-			/// <returns></returns>
-			public string GetRepresentation (object value)
-			{
-				if (!string.IsNullOrWhiteSpace (Format)) {
-
-					if (value is IFormattable f)
-						return f.ToString (Format, null);
-				}
-
-				if (RepresentationGetter != null)
-					return RepresentationGetter (value);
-
-				return value?.ToString ();
-			}
-		}
-		/// <summary>
-		/// Defines rendering options that affect how the table is displayed.
-		/// 
-		/// <a href="https://gui-cs.github.io/Terminal.Gui/articles/tableview.html">See TableView Deep Dive for more information</a>.
-		/// </summary>
-		public class TableStyle {
-
-			/// <summary>
-			/// Gets or sets a flag indicating whether to render headers of a <see cref="TableView"/>.
-			/// Defaults to <see langword="true"/>.
-			/// </summary>
-			/// <remarks><see cref="ShowHorizontalHeaderOverline"/>, <see cref="ShowHorizontalHeaderUnderline"/> etc
-			/// may still be used even if <see cref="ShowHeaders"/> is <see langword="false"/>.</remarks>
-			public bool ShowHeaders { get; set; } = true;
-
-			/// <summary>
-			/// When scrolling down always lock the column headers in place as the first row of the table
-			/// </summary>
-			public bool AlwaysShowHeaders { get; set; } = false;
-
-			/// <summary>
-			/// True to render a solid line above the headers
-			/// </summary>
-			public bool ShowHorizontalHeaderOverline { get; set; } = true;
-
-			/// <summary>
-			/// True to render a solid line under the headers
-			/// </summary>
-			public bool ShowHorizontalHeaderUnderline { get; set; } = true;
-
-			/// <summary>
-			/// True to render a solid line vertical line between cells
-			/// </summary>
-			public bool ShowVerticalCellLines { get; set; } = true;
-
-			/// <summary>
-			/// True to render a solid line vertical line between headers
-			/// </summary>
-			public bool ShowVerticalHeaderLines { get; set; } = true;
-
-			/// <summary>
-			/// True to render a arrows on the right/left of the table when 
-			/// there are more column(s) that can be scrolled to.  Requires
-			/// <see cref="ShowHorizontalHeaderUnderline"/> to be true.
-			/// Defaults to true
-			/// </summary>
-			public bool ShowHorizontalScrollIndicators { get; set; } = true;
-
-
-			/// <summary>
-			/// Gets or sets a flag indicating whether there should be a horizontal line after all the data
-			/// in the table. Defaults to <see langword="false"/>.
-			/// </summary>
-			public bool ShowHorizontalBottomline { get; set; } = false;
-
-			/// <summary>
-			/// True to invert the colors of the first symbol of the selected cell in the <see cref="TableView"/>.
-			/// This gives the appearance of a cursor for when the <see cref="ConsoleDriver"/> doesn't otherwise show
-			/// this
-			/// </summary>
-			public bool InvertSelectedCellFirstCharacter { get; set; } = false;
-
-			/// <summary>
-			/// Gets or sets a flag indicating whether to force <see cref="ColorScheme.Normal"/> use when rendering
-			/// vertical cell lines (even when <see cref="FullRowSelect"/> is on).
-			/// </summary>
-			public bool AlwaysUseNormalColorForVerticalCellLines { get; set; } = false;
-
-			/// <summary>
-			/// Collection of columns for which you want special rendering (e.g. custom column lengths, text alignment etc)
-			/// </summary>
-			public Dictionary<int, ColumnStyle> ColumnStyles { get; set; } = new Dictionary<int, ColumnStyle> ();
-
-			/// <summary>
-			/// Delegate for coloring specific rows in a different color.  For cell color <see cref="ColumnStyle.ColorGetter"/>
-			/// </summary>
-			/// <value></value>
-			public RowColorGetterDelegate RowColorGetter { get; set; }
-
-			/// <summary>
-			/// Determines rendering when the last column in the table is visible but it's
-			/// content or <see cref="ColumnStyle.MaxWidth"/> is less than the remaining 
-			/// space in the control.  True (the default) will expand the column to fill
-			/// the remaining bounds of the control.  False will draw a column ending line
-			/// and leave a blank column that cannot be selected in the remaining space.  
-			/// </summary>
-			/// <value></value>
-			public bool ExpandLastColumn { get; set; } = true;
-
-			/// <summary>
-			/// <para>
-			/// Determines how <see cref="TableView.ColumnOffset"/> is updated when scrolling
-			/// right off the end of the currently visible area.
-			/// </para>
-			/// <para>
-			/// If true then when scrolling right the scroll offset is increased the minimum required to show
-			/// the new column.  This may be slow if you have an incredibly large number of columns in
-			/// your table and/or slow <see cref="ColumnStyle.RepresentationGetter"/> implementations
-			/// </para>
-			/// <para>
-			/// If false then scroll offset is set to the currently selected column (i.e. PageRight).
-			/// </para>
-			/// </summary>
-			public bool SmoothHorizontalScrolling { get; set; } = true;
-
-			/// <summary>
-			/// Returns the entry from <see cref="ColumnStyles"/> for the given <paramref name="col"/> or null if no custom styling is defined for it
-			/// </summary>
-			/// <param name="col"></param>
-			/// <returns></returns>
-			public ColumnStyle GetColumnStyleIfAny (int col)
-			{
-				return ColumnStyles.TryGetValue (col, out ColumnStyle result) ? result : null;
-			}
-
-			/// <summary>
-			/// Returns an existing <see cref="ColumnStyle"/> for the given <paramref name="col"/> or creates a new one with default options
-			/// </summary>
-			/// <param name="col"></param>
-			/// <returns></returns>
-			public ColumnStyle GetOrCreateColumnStyle (int col)
-			{
-				if (!ColumnStyles.ContainsKey (col))
-					ColumnStyles.Add (col, new ColumnStyle ());
-
-				return ColumnStyles [col];
-			}
-		}
+		
 
 		/// <summary>
 		/// Describes a desire to render a column at a given horizontal position in the UI
@@ -2045,41 +1817,5 @@ namespace Terminal.Gui {
 			}
 
 		}
-
-		/// <summary>
-		/// Describes a selected region of the table
-		/// </summary>
-		public class TableSelection {
-
-			/// <summary>
-			/// Corner of the <see cref="Rect"/> where selection began
-			/// </summary>
-			/// <value></value>
-			public Point Origin { get; set; }
-
-			/// <summary>
-			/// Area selected
-			/// </summary>
-			/// <value></value>
-			public Rect Rect { get; set; }
-
-			/// <summary>
-			/// True if the selection was made through <see cref="Command.ToggleChecked"/>
-			/// and therefore should persist even through keyboard navigation.
-			/// </summary>
-			public bool IsToggled { get; set; }
-
-			/// <summary>
-			/// Creates a new selected area starting at the origin corner and covering the provided rectangular area
-			/// </summary>
-			/// <param name="origin"></param>
-			/// <param name="rect"></param>
-			public TableSelection (Point origin, Rect rect)
-			{
-				Origin = origin;
-				Rect = rect;
-			}
-		}
-		#endregion
 	}
 }
