+++ UnitTests/Views/TableViewTests.cs
@@ -327,7 +327,7 @@ namespace Terminal.Gui.ViewsTests {
 
 			// select the last row
 			tableView.MultiSelectedRegions.Clear ();
-			tableView.MultiSelectedRegions.Push (new TableView.TableSelection (new Point (0, 3), new Rect (0, 3, 4, 1)));
+			tableView.MultiSelectedRegions.Push (new TableSelection (new Point (0, 3), new Rect (0, 3, 4, 1)));
 
 			Assert.Equal (4, tableView.GetAllSelectedCells ().Count ());
 
@@ -430,8 +430,8 @@ namespace Terminal.Gui.ViewsTests {
 			*/
 
 			tableView.MultiSelectedRegions.Clear ();
-			tableView.MultiSelectedRegions.Push (new TableView.TableSelection (new Point (1, 1), new Rect (1, 1, 2, 2)));
-			tableView.MultiSelectedRegions.Push (new TableView.TableSelection (new Point (7, 3), new Rect (7, 3, 2, 1)));
+			tableView.MultiSelectedRegions.Push (new TableSelection (new Point (1, 1), new Rect (1, 1, 2, 2)));
+			tableView.MultiSelectedRegions.Push (new TableSelection (new Point (7, 3), new Rect (7, 3, 2, 1)));
 
 			tableView.SelectedColumn = 8;
 			tableView.SelectedRow = 3;
@@ -1534,7 +1534,7 @@ namespace Terminal.Gui.ViewsTests {
 
 			// user has rectangular selection 
 			tableView.MultiSelectedRegions.Push (
-				new TableView.TableSelection (
+				new TableSelection (
 				new Point (0, 0),
 				new Rect (0, 0, 3, 1))
 				);
